+++ lib/rubocop/cop/lint/block_alignment.rb
@@ -12,101 +12,38 @@ module RuboCop
       #     i
       #   end
       class BlockAlignment < Cop
-        include CheckAssignment
-
         MSG = '`end` at %d, %d is not aligned with `%s` at %d, %d%s.'
 
         def on_block(node)
-          return if ignored_node?(node)
           check_block_alignment(start_for_block_node(node), node)
-        end
-
-        def on_and(node)
-          return if ignored_node?(node)
-
-          _left, right = *node
-          return unless right.type == :block
-
-          check_block_alignment(node, right)
-          ignore_node(right)
-        end
-
-        alias_method :on_or, :on_and
-
-        def on_op_asgn(node)
-          variable, _op, args = *node
-          check_assignment(variable, args)
-        end
-
-        def on_send(node)
-          _receiver, _method, *args = *node
-          check_assignment(node, args.last)
-        end
-
-        def on_masgn(node)
-          variables, args = *node
-          check_assignment(variables, args)
         end
 
         private
 
-        def check_assignment(begin_node, other_node)
-          return unless other_node
-
-          block_node = find_block_node(other_node)
-          return unless block_node.type == :block
-
-          # If the block is an argument in a function call, align end with
-          # the block itself, and not with the function.
-          if begin_node.type == :send
-            _receiver, method, *_args = *begin_node
-            begin_node = block_node if method.to_s =~ /^\w+$/
-          end
-
-          # Align with the expression that is on the same line
-          # where the block is defined
-          if begin_node.type != :mlhs && block_is_on_next_line?(begin_node,
-                                                                block_node)
-            return
-          end
-          return if ignored_node?(block_node)
-
-          ignore_node(block_node)
-          check_block_alignment(begin_node, block_node)
-        end
-
-        def find_block_node(node)
-          while [:send, :lvasgn].include?(node.type)
-            n = case node.type
-                when :send
-                  find_block_or_send_node(node) || break
-                when :lvasgn
-                  _variable, value = *node
-                  value
-                end
-            node = n if n
-          end
-          node
-        end
-
-        def find_block_or_send_node(send_node)
-          receiver, _method, args = *send_node
-          [receiver, args].find do |subnode|
-            subnode && [:block, :send].include?(subnode.type)
-          end
-        end
+        def_node_matcher :block_end_align_target?, <<-PATTERN
+          {assignment?
+           splat
+           and
+           or
+           (send _ :<<  ...)
+           (send equal?(%1) !:[] ...)}
+        PATTERN
 
         def start_for_block_node(block_node)
           # Which node should we align the 'end' with?
           result = block_node
 
-          while result.parent &&
-                (result.parent.splat_type? ||
-                (result.parent.send_type? &&
-                  !operator?(result.parent.method_name)))
-            result = result.parent
+          while (parent = result.parent)
+            break if !parent || !parent.loc
+            break if parent.loc.line != block_node.loc.line &&
+                     !parent.masgn_type?
+            break unless block_end_align_target?(parent, result)
+            result = parent
           end
 
+          # In offense message, we want to show the assignment LHS rather than
+          # the entire assignment
+          result, = *result while result.op_asgn_type? || result.masgn_type?
           result
         end
 
@@ -146,12 +83,8 @@ module RuboCop
           end
         end
 
-        def block_is_on_next_line?(begin_node, block_node)
-          begin_node.loc.line != block_node.loc.line
-        end
-
         def autocorrect(node)
-          ancestor_node = ancestor_on_same_line(node)
+          ancestor_node = start_for_block_node(node)
           source = node.loc.expression.source_buffer
 
           lambda do |corrector|
@@ -168,13 +101,6 @@ module RuboCop
                                                 starting_position_of_block_end)
               corrector.remove(range)
             end
-          end
-        end
-
-        def ancestor_on_same_line(node)
-          node.ancestors.reverse.find do |ancestor|
-            next unless ancestor.loc.respond_to?(:line)
-            ancestor.loc.line == node.loc.line
           end
         end
       end